classes: add qa-denydeps
This class adds functionality to run a test that checks an RDENYDEPENDS
variable for packages that are forbidden to be a transitive dependency
of a package. In particular, this can be used to ensure that certain
troublesome packages do not get pulled in via certain packagegroups,
while still allowing said packages to be built and installed outside
of the packagegroup.

Signed-off-by: Brandon Streiff <brandon.streiff@ni.com>