packagegroup-ni-base: ensure python3 stays out of base
The base packagegroup should contain the bare minimum set of packages
for all NI use cases, and some of those use cases are very tightly
constrained, such as ARM safemode images. The python3 runtime is
large and it keeps finding its way accidentally into places, so add
a specific QA check for it to alert us when it happens.

Signed-off-by: Brandon Streiff <brandon.streiff@ni.com>